Move logic to fragment.
